% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{schema_path}
\alias{schema_path}
\title{Return the path to the vendored schema.json on disk.}
\usage{
schema_path()
}
\value{
Length-1 character vector pointing at the installed schema.json.
}
\description{
Return the path to the vendored schema.json on disk.
}
\keyword{internal}
